Integrated GEM5 changes into Aladdin.
This change merges all of Sophia's past changes from the gem5-aladdin branch
into the main Aladdin repository and brings some of the older changes in sync
with the newer changes to Aladdin.

New directory structure:
  * General Aladdin components now belong inside common/
  * Simulator specific modifications belong inside a separate folder.

Original Makefile moved to common/. SConscripts are GEM5 specific and are in
gem5/.

GEM5 and Aladdin both build successfully with this merge and pass (one) test!

Change-Id: Iff2b4cac63b96ba29b0ada6ad4aaf6331190afab