Fixed the issue of loop pipelining when Aladdin tried to pipeline different functions. Loop pipelining should only applied to loops, not functions. This fix added dependences between two consecutive branch ops, including call ops, and do not remove dependences between branches when the dependent child is a call op.
Somehow the globalOptimizationPass() did not include
memoryAmbiguation()...Added it back. The idea of the memoryAmbiguation() pass is
to add memory dependences to the load store parirs that can not be easily
statically disambiguated. In our case, if a load-store dependence shows up in
one instance, we assume this applies to all the dynamic instances of the same
load-store pairs.

Change-Id: I7367066bcd95edded8e87cfa1a5e0a674113e8f8