Support for multiple accelerators that depend on other accelerators.
This is accomplished by assigning each accelerator an ID and a list of
dependencies. Accelerators with dependencies must wait for all dependencies to
complete execution before their execution can start. Currently, there is zero
latency between the completion of one accelerator and the start of a dependent
accelerator, but we can introduce an arbitrary amount of latency very easily.

Change-Id: I8cfc23f685ef0717af615dd39334d30a887f473b