Preliminary support for writing results to MySQL database.
From this point forward, the mysql c++ connector libraries and include headers
are required. Specify the location of the include headers as the environment
variable MYSQL_HOME and add the shared library locations to LD_LIBRARY_PATH.
Database access parameters are specified in DatabaseConfig.h

Currently, only ScratchpadDatapath has been added with this support, so gem5 may
not yet compile.

To use, invoke the aladdin binary as usual, but add an extra experiment_name
parameter at the end. This parameter will be used to identify a set of related
experiments.

BUG=ALADDIN-20

Change-Id: Id1fe83bfa6a5d6983128640a621b6329cfd14f3c