Change the way number of multipliers is modeled in Aladdin. There are two ways of model number of functional units, one is to count the maximum number of functional units per cycle and assume that the functional units can be perfectly shared across cycles; the other is to count the number of computational operators in an unrolled loop region and assuem that there is no sharing within an unrolled region and unrolling means exactly duplication.
The first one is optimistic in terms of wiring cost when it comes to sharing,
and the second one is too pessimistic.

When we compare these assumptions with Vivado generated RTL, we found that the
number of multipliers is more close to the assumption one and the number of
adders is modeled similar to assumption two.

The reason for that could be the relative cost of functional units with respect
to wires. Though wiring cost may occur when it comes to sharing, the multiplier
cost is more significant. That's why we try to share multipliers whenever it's
possible. While for adders or even smaller bit-wise operators, the wiring cost
would be even more expensive than duplication of functional units, which
explains the adders are less likely to be shared across iteration boundaries.

Change-Id: I2aca038edfe67b2031a8e2c46330787b06d175a1