Support for invoking DmaScratchpadDatapath through user program.
In syscall emulation mode, gem5 doesn't support the use of an iobus that is
attached as a secondary bus to the main memory bus. Typically, IO devices are
made cache coherent through an IO cache on the IO-bus. To workaround this, I've
added a tiny cache attached to the datapath on a new port. Typical DMA requests
use a different port and thereby bypass the cache entirely, but synchronization
messages that need to be made coherent will go through this little cache.

The interface for invoking the accelerator doesn't change.

Change-Id: Iae4627c4b2046a842f740e17badf6eb2ca4596e7