Converting the ioctl and mapping utility functions from C++ to C.
These are also being renamed to reflect that they are not just for ioctl
syscalls, but generally for enabling connections to the rest of the system.

Conversion to C will enable these headers to be used by MachSuite, SHOC, and any
other C benchmark without changing compilers. They are specified as .cpp files
so gem5 can compile and link with g++, but they can also be compiled with C code
using gcc as well.

BUG=ALADDIN-24

Change-Id: I6bbb34f52ccb640487d443cc05d67a784e7ab018