Update the TLB and cache models so that they use real translations.
Formerly, the physical page number returned by the TLB was never actually used,
and the cache was accessed using the array's virtual address. Now the
virtual address is translated to a physical address and that translation is used
to access the cache.

If entries in the TLB are evicted, they are stored in a map, and TLB misses will
consult this map to get the new mapping. Later we will push this functionality
into the page table model itself.

Still outstanding: TLB misses can be caused by accesses to completed partitioned
arrays, when instead they should not have to go through the cache hierarchy at
all.

BUG=ALADDIN=25

Change-Id: I49efb0c1c2d6a68643e6d52d80fb2937c8340735