Fixed an issue shown up in bb_gemm. removeRepeatedStores() was not working because it checks whether a store node has any children, including the control edges. To fix it, we specifically check whether a store has any non-control children. This fix removed the repeated stores shown up in bb_gemm.
Also switched executingQueue and readyToExecuteQueue from vector to list. The
reason here is that we want to execute ready stores as soon as possible. So when
we insert stores into the executingQueue, we want the stores to be at the top,
i.e., push_front instead of push_back().

Also fixed a minor issue in initializing executingQueue(). We use the number of
nodes in the graph, i.e., nameToVertex.size(), instead of the total number of
nodes in the trace, i.e., numTotalNodes. The reason here is that we could have
slightly fewer number of nodes in the graph if the nodes at the end of the trace
are not connected at all. Using numTotalNodes in setGraphForStepping() may add
wrong mapping between node names and node vertex into nameToVertex.

Change-Id: I1f143907736cf866eabf3fabeefc94b3a59427cb