Fixed an issue with chained phi nodes, shown up in benchmark md-grid.
The original removePhiNodes() removes edges around phi nodes, then add
edges between phi nodes' children and parents. This won't work when we have a
chain of phi nodes, since edges will be added between a phi node's parent and
children, in which case its parent is still a phi node.

To fix this, we added a loop in searching for phi nodes' parents, the loop
doesn't stop until we find a non-phi node parent, then update edges between
phi's non-phi parent and children.

Notice that phi node can only have one parent, although it takes two parameters.
Depending on what the previous basic block is, it takes the value associated
with the previous basic block. To consider that, we also modified DDDG.cpp to
ignore the parameter line if it does not match the previous basic block.

This fix requires a pull of LLVM-Tracer. We modified LLVM-Tracer to profile one
more column for phi nodes, the previous basic block id associated with each
parameter.

Updated unit-test trace files for triad-128, triad-initbase, and reduction.

Change-Id: Ie8f733fafa4e44acb86fbcb459e804b885e0eccc