Fix address translation bug for user-mapped arrays.
The problem was that we were assuming that the dynamic addresses found in the
trace are page aligned in the same way that the simulated address space is.

Example:

In the trace, we see the address 0xAA00 is the base address of an array.  The
simulated virtual address is 0xBBE0, which gets translated to a physical address
like 0xFFE0. So, the page translation is 0xBB to 0xCC. The existing
implementation adds a new mapping 0xAA to 0xCC, but that results in an incorrect
translation because the offsets are different. The trace's offset is 0x00, but
the simulated offset is 0xE0. Instead of translating 0xAA00 to 0xFFE0, we would
translate to 0xFF00. This results in the stores being committed to the wrong
address.

We fix this by changing the translation scheme. Instead of going straight from
trace to physical address, we first translate the trace base address to the
virtual base address, account for the differences in offsets, and then
translate from simulated virtual to physical. Therefore, the TLB now ONLY stores
the simulated virtual addresses, which is better since that's all it's supposed
to store anyways.

BUG=ALADDIN-25

Change-Id: I1271a55a3c053cfa585aafb46c889c5ee1208da7