First integration test for gem5-aladdin!
Each additional test should have its own folder and Makefile that specifies the
list of source files and other metadata as shown in this Makefile. The makefile
will generate the instrumented binary and generate the dynamic trace.

This test ensures that stores performed by the accelerator are visible to the
CPU.

Change-Id: Iccdca31c92876619e05f97352a6fbe1f9a6f2d40