Relax the condition in memoryAmbiguation().
Using dynamic function id and instruction id is sufficient as a unique
identifier for load-store pairs. Before we also used previous basic block id as
part of the identifier, which was causing problems when we have the same load
store pairs but from different basic blocks, like the first iteration of a loop
and the rest iterations, then we were double-counting dependences.

Change-Id: If6c5c602192662348d4c86dd1c9ea74d6d8219ae