Speedup Aladdin simulation by removing redundant scratchpad checking.
We observed that Aladdin simulation is slower for certain benchmarks when
pipelining is enabled than disabled. The reason for that is when Aladdin does
stepExecutingQueue() in ScratchpadDatapath, for every memory operation in
executingQueue, we check whether scratchpad can service or not. This can lead to
redundant checking when the entire scrachpads cannot service any more but we
still check canServicePartition(node_part). The reason that pipelined config
leads to more slowdown compared to unpipelined is that pipelined design is more
likely to have more nodes in the executingQueue due to parallelism, which leads
to more redundant checking compared to unpipelined designs.

To fix this, we add a scratchpadCanService flag inside ScratchpadDatapath. At
each cycle, if we find all the scratchpads can not service any more, we skip the
memory operations, no more checking.

Also fixed Scratchpad occupiedBWPerPartition checking. Was not updated when we
increment loads and stores...

Change-Id: Ib1218d0218633d53c4099ccb84bbd2e249585297