Refactor some of the accelerator invocation code and add a non-blocking option.
This means that client code does not need to set up the finish flags. The code
instead has an option for a blocking invocation or a nonblocking invocation.
The non-blocking option returns the pointer to the finish flag after invoking
the accelerator, and it is up to the client code to check for accelerator
completion.

Change-Id: Id7b55ce6d86696345ebb37703834f87708add70f