Update Aladdin's FPGA model.
We update Aladdin's ScratchpadDatapath to model FPGA where multiple functional
units could be packed into a single cycle. We use the functional units latency
from Virtex7 board.

Updated removeInductionDependence() to mark nodes that are inductive if all
their parents are inductive.

Also fixed an issue in removePhiNodes(): inside the loop to find phi nodes
parents, it's possible that the final phi node doesn't have a parent.

Also simplified to memoryAmbiguation() pass to leverage the induction_nodes()
information we get from removeInductionVariables(): if the gep node for memory
operations are induction nodes, the memory ops don't need to be ambiguaged.

Added a field in summary for max_reg.

Passed unit tests.

Change-Id: I9d837fc1e8b2fd7bda32b6c1f38639080011fb57