Restore the isolated flag to BaseNode to fix initExecutingQueue().
Bug symptom: Aladdin would hang on "Scheduling".
Bug cause: a node is isolated if it has both no parents AND no children. We
previously only checked if it had no parents. This caused nearly all of the
nodes to be placed on the executing queue in the first cycle, instead of just
the ones with no parents.

Also, change execution_cycle to int from unsigned int to fix a bug in
clearGraph, where execution_cycle was getting assigned a value of -1.

Change-Id: Ife97149cb8f1d2fc9487cbf59e30943ac503739d