A datapath implementation that supports both caches and scratchpads.
Most of the code came verbatim from DmaScratchpadDatapath and CacheDatapath,
but I've cleaned up a lot of redundant functions and APIs. The datapath can be
used in cache-only mode or in spad-only mode by appropriately specifying the
configuration parameters in the benchmark configuration:

cache,[array_name],[array_size],[word_size]

The current implementation completes the new integration test successfully and
I've done a rough validation of the performance statistics seems okay, but I
have not validated the power calcuations at all.

All integration tests pass except for test_cortexsuite, due to a bug unrelated
to this port, so I will address that later.

CacheDatapath will not compile anymore since I've changed things in
aladdin_tlb, so I've taken it out of the SConscript. We'll delete the file
later completely.

BUG=ALADDIN-27

Change-Id: I0e8a4b627a9e714b5c56a7d5eb1f7bec3531c515