Fixed aes crash in the standalone simulation.
Although gem5's X86_64 implementation supports 48-bit address space, in
accelerator standablone mode, the membus address range is the same as the memory
size, which is 2GB now. In this case, addresses longer than 31-bit cannot be found
in the membus, which causes the crash.

We added an ADDR_MASK to mask out the higher order bits in the address for
standalone simulation mode. Currently we assume having a 4GB memory. Will also update
the sweep script to use 4GB memory size by default.

Change-Id: I8e68582e1cc17881ded7e88b241858c284779066