Fixed the unknown address error in initBaseAddress() for aes.
The error happens because there is not GetElementPtr instruction before
instructions that load/store from rcon. LLVM performs instruction combination
(the instcombine pass) to combine gep and load/store for rcon, which is a
one-element pointer in the aes case.

To fix it, we update last_written filed for the register that stores the
load/store address, rcon in this case, when load/store happens, treating it as a
combination of gep and load/store, which writes the address register and does
the memory operations at the same time.

However, this fix has its limitation. It fixes aes for now because
luckily aes has a store rcon before it calls aes_expandEncKey() where rcon turns
into rc. If there is no store to rcon before the call, Aladdin cannot trace the
pointer rc in aes_expandEncKey() back to the pointer rcon in the parent
function. Working on that now.

Change-Id: I937a2f29dee83f1c70b823532b09211a90d8d33a