Fixed integration test in initBaseAddress() with both spad and Hybrid.
The reason it fails is that we only perform initBaseAddress for connnected nodes
in the graph, using the graph vertex iterator, but check whether base address is
initialized for all the mem nodes in exec_nodes, which has more nodes than
connected nodes in the graph.

In the case of integration, there is a load operation, the value of which is
used in a malloc operation that is not traced. So it's a standalone load, which
is eliminated in the cleanLeafNodes().

The fix also uses the graph vertex iterator to check all the connected nodes,
which is the same as how we perform initBaseAddress().

Change-Id: I9bd1f651b34a856a74098493979be9e8a48cfc93