Use FU critical path delay as FU's timing model.
Originally we use FPGA FU latency characterization in packing since we are
validating against Vivado which is an FPGA flow. However, there is a mismatch
between the FU latency from FPGA and the power model from ASIC flow.

In this commit, we use the critical path delay from Design Compiler's synthesis
results to accurately model FU's timing. Also since the ScratchpadDatapath class
is used for HybridDatapath as well, the memory latency is not fixed anymore. We
change the node_latency in ExecNode to fu_node_latency since it only gives
non-memory node latency.

Change-Id: I96497283a9597f0c467326e20dbb056ef0c80818