Added clearDatapath() to clear states for multiple invocations.
Added a file descriptor inside BaseDatapath to keep track of how much the trace
file has scanned so far. DDDG only builds the dddg for one invocation of a
function. When the function is called again, in a different invocation, DDDG
start scanning the trace from where it was left and builds a new DDDG until this
innocation ends (returns).

clearDatapath() is added to clear the relevant states at the end of each
invocation simulation. Also added clear() function to relevant classes, like
Registers and Scratchpad.

Change-Id: I087cf88933b1d7fb8fdf4ebf695261ef36f683cf