Add option to dump gem5 stats after each accelerator invocation.
This means we can now get the entire simulator statistics per invocation, not
just the ones that we're tracking ourselves, which is a big win. This is
accomplished by calling exitSimLoop() at the end of each invocation. The
simulation script now looks for the cause of simulation exit, and if it's to
dump stats, then after doing so, it will resume simulation where it ended last.

This will require an associated change in gem5.

Change-Id: Ib51d859c80fac010a4a558063d47263c800641c5