Simplify request code valid check.
Instead of checking if the code is within a valid range AND it does not
belong to a special value, just check the latter. We always forget to
update the valid range anyways (ideally this would be done by default),
and even that doesn't really guarantee that the code is really valid (it
IS assigned to an accelerator that EXISTS and is REGISTERED with the
system).

Change-Id: Ia01f555016e81e9067bca56746c96693ddc6a442