Use the base address as an unique identifier for dma ops.
The array label of DMA nodes is initialized in DDDG, where it can only see the
label specified in the current instruction. We do not have an initBaseAddress()
for dmaLoad and dmaStore, so the labels inialized in DDDG() won't be updated
with dependence analysis. In this case, we may run into labels that are not in
the partition config.

Instead of adding dma nodes handling in initBaseAddress(), we use base address
as a unique identifier for dma ops for now. It's a hacky way to find the base
address of DMA operations, should be fixed later.

Change-Id: I7081191def03fd7179ab3e58d1008ebd836dfea6