Stricter checking of matching branches for loop pipelining.
We only want to pipeline loop iterations that are from the unrolled loop. Before
we were not checking whether they are from the same loop explictly, because even
if we pipeline two loop iterations from different loops, it only affects the
boundaries between different loops, as in the case between the last iteration of
the previous loop and the first iteration of the next loop.
But it has bigger impact for dma operations that if we pipeline a dma operation
and a loop iteration, the dma operation will not be blocking the later
computation.

To fix this issue, we add a checker before loop pipelining to make sure that we
only pipeline the iterations from the same loop. As I'm working on this, I also
make the loop unrolling counter starts from 0 instead of 1, as when we have a
unrolling factor of 2, we want iteration 0 and 1 in the same unrolled iteration,
then 2 and 3, and so on.

These two changes to loop pipelining and unrolling also impact unit tests. The
checkers in unit tests also updated accordingly.

Change-Id: I330070e61d7d9a2115e3ca5c8b6eeafd55b33f44