Refactored Scratchpad class in preparation for the ready-bit change.
Added two new classes: LogicalArray and Partition.

A LogicalArray stores meta data of a program defined array, like total size,
word size, partition type, but it does not store the values as well as status
bits that we are going to add later. Based on the partition factor, it can
contain multiple Partitions. Certain sharable information across Partitoins is
also stored here, like per access read/write energy, leakage power, and area,
all of which are the same across partitions for the same array.

A Partition describes each partitioned array. It keeps track of how many
accesses that have happened to the partitioned array. In the runtime, it also
checks whether there is enough bandwidth, or whether specific tag bit is set
(ready-bit) in the future, to decide whether it can take a memory request or
not.

After this refactoring, when Datapath sends scratchpad acceess requests to
Scratchpad, instead of generating a compound string, like "a-0", it sends the
base array label, e.g., "a", and the partition index, e.g., 0, to Scratchpad
directly. Scratchpad is responsible to map the request to the specific Partition
to see whether it can service at the current cycle or not.

BUG=ALADDIN-47

Change-Id: I07e971d72f0c12c3c51080664a3fc8d17edc5c7b