Use operation-specific latencies for FP ops.
Previously, all FP ops took 3 cycles. This is clearly not true for
divides, and adds tend to take more time than multiplies. We now use
different latencies for adds, muls, and divs, taken from the Zynq
Zedboard.

To be even more accurate, we can have different latencies for different
clock frequencies, but we'll defer that for now.

Change-Id: Ia65ed4a582dce5fb1e023bd0184279f38931aa38