Add cache line flush and invalidate latency to DMA.
These two latencies together form the complete DMA setup latency. Based
on a parameter, we can either force all DMA nodes to wait for EVERYONE
to finish the setup time, or we can overlap setup and transfer of
individual nodes.

For now, the values of these latencies are provided as hardcoded
constants in HybridDatapath.  I don't think we will need to or ever want
to sweep this as a parameter.  We should find a new home for these
constants eventually though.

This also make the default DMA setup latency more reasonable (1000
accelerator cycles is way too high), and fix a bug where the
dmaSetupLatency was interpreted as Ticks instead of Cycles.

Change-Id: Ic37b042e95d040e6cc55edb46b27a813b5fc9684