Pipelined DMA and flush implementation.
Required manual changes to all benchmarks as well as the addition of an
offset field to the DMA interface and additional tracking of base
addresses for handling responses (for ready bits).

Assertions have been added to the ready bits to help catch segfaults,
and a histogram stat that we don't use is removed.

Change-Id: I73646c57c619643924cd31aea81e6de25617675d