bfs: Fix deadlock due to one array not dmaLoaded.
Because level_counts is so small (80 bytes), we don't dmaLoad it right
now (as the overhead would be huge per byte). However, its ready bits
never get set by a different ordinary store before an ordinary load,
leading to a deadlock when trying to load from level_counts.

We don't want to dmaLoad level_counts because it's just 80 bytes, and
the current DMA system would require a flush and invalidate of 80 bytes,
which is very expensive. The correct way would be dmaLoad an entire
chunk of memory without any regard to array boundaries. For now, just
initialize level_counts to zero - this will flip the ready bits with
minimal overhead.

Change-Id: I10b319a93e7c97f974ae5614311d0bb0a09450db