Make the MemoryQueue a real hardware tracking structure.
This reintroduces size and bandwidth constraints and power modeling for
this queue (turns out taking this all away was not such a good idea
after all). This combines loads and stores together into a single
bandwidth (for simplicity).

The integration test test_load_store has been updated to test this
behavior by unrolling the inner loop and but limiting the performance by
setting cache bandwidth to be much less than the unrolling factor.

BUG=ALADDIN-51
BUG=ALADDIN-68

Change-Id: I62bdd5734ea2c2a4c5195a09c51f37b8f1210c59