dma: Introduce version 2 of DMA interface.
This version separates the offset field into source and destination
offsets. This means that an array can request for any offset from the
base address from main memory and store it at any other valid offset in
the local scratchpad, and vice versa when storing data back to main
memory.  This is required for double buffering.

In addition to separated offsets, the DMA loads and stores need to
actually move data on the host so that the trace will contain the
correct values for loads and stores.

The v1 interface is now DEPRECATED; it will remain as the default for
the time being but will be removed in the future. Users who want to use
the v2 interface for now must define the macro DMA_INTERFACE_V2 in
CFLAGS.

Change-Id: I1b88ac6b95fac4f6fe37117e3c223e768d2df73d