dma: Introduce DMA fences.
DMA operations are not ordered with respect to one another, only via
memory dependences to ordinary loads and stores. This design is because
DMA is typically handled as a separate "thread" or hardware module, and
we DO often want multiple DMA operations to proceed at the same time.
DMA fences provide a simple way for the programmer to enforce ordering
when it is required. Like memory fences, all DMA operations must commit
before the fence will commit, and all subsequent DMA operations must
wait for the fence to commit before they can issue.

Change-Id: I717152c79ced9ac9e8412f61ac305001195679bc