base: Add debugging functions to SourceManager.
Without these functions, it's pretty difficult to debug, particularly as
calling templated functions in a debugger isn't so pretty.  This also
fixes some bugs in SourceManager relating to invalid source entity
objects.

To use in gdb, go to a stack frame under BaseDatapath:
  $ call srcManager.dump(<insert id>)

Passing around const references is fine and all but checking for
invalidity is becoming a huge pain. I should refactor and replace all
references with pointers, so that a NULL check is all that's needed, but
that might take awhile.

Change-Id: I0c50cd6245ce948b1aca3b182477692464a279dc