gem5: Add a sim_cycles stat.
This mirrors the num_cycles counter in the BaseDatapath. Adding this as
a stat means it will be automatically dumped in the set of statistics,
making it easier to parse.

Update all integration tests to check the accelerator sim cycles instead
of the system sim_ticks, because sim_ticks is prone to flakiness due to
differences in host environments.

Change-Id: Iad0e1ebb2f7aeef7b735f6eb9063573c9a6f3f44