common: Implementation of per-loop pipelining.
This addresses several limitations of the previous global loop
pipelining design, although the implementation borrows a lot of its
code.

  1. We can select specific loops to pipeline (rather than all or
     nothing).
  2. Loops with a non-flattened subloop of iteration 1 can still be
     pipelined. Example:

     int N = 3;  // this can be anything.
     int M = 1;  // must be equal to 1.
     outer: for (int i = 0; i < N; i++) {
       middle: for (int j = 0; j < M; j++) {
         inner: for (int k = 0; k < 8; k++) {
           /* do work */
         }
       }
     }

     If inner is flattened, we can pipeline middle and have a new
     iteration of middle start before the previous one is finished
     (assuming no other loop-carried dependencies exist). This would be
     almost equivalent to pipelining outer as long as middle only runs
     once.

The current primary shortcoming of this implementation is the inability
to recognize when a loop cannot be pipelined. In the above example, if
M > 1 and we pipeline outer (but not middle), what happens is that M
iterations of middle will execute serially, but they will be pipelined
with the NEXT group of M iterations of middle from the next iteration of
outer. Given that true memory dependences probably exist between
iterations of the outer loop, this will not actually allow Aladdin to
find much inter-iteration parallelism.

Change-Id: Ic89d77032cab0d1c05ca1f5f3b8aeb424a4094e2