common: Replace references with pointers in SourceManager.
References seemed like a good idea originally: to avoid null pointer
issues, we'd just store their IDs and compare them instead, and we'd use
the source manager if we actually needed more information. But now
they're just a huge pain because we ended up with inconsistencies in how
we use Invalid<Type>, invalid ids, etc, and the SourceManager itself had
to be passed around everywhere.

Now, we store the raw pointers themselves, which means we don't need to
refer to the SourceManager nearly as often, and null pointers represent
invalid objects. For static SourceEntities, we can compare the raw
pointers directly (instead of the operator==), because now only the
SourceManager is allowed to construct SourceEntity objects, and it will
always return a pointer to an existing object with a given name before
constructing a new one.  Dynamic entities are meant to be constructed on
the fly, so we can't just compare pointers.

Change-Id: Ic78a9a4111c4c3e4e15ab694618fa3a4215d9f09