common: Fuse register load/stores with surrounding ops.
A "load" typically implies that an array is being read and its data
stored in a temporary register, but if that value is directly being used
by a compute operation or stored to another register, the temporary
register (and additional one-cycle latency incurred by the load) is
extraneous.

This optimization fuses register loads and stores with the next or
previous operation, respectively, by creating a register dependence edge
between the two nodes. The scheduler will allow two nodes connected by a
register dependence to execute in the same cycle.

This sort of optimization can more accurately model the behavior of
compute patterns like multiply-accumulate operations and shift
registers, both of which often involve repeated operations on values
stored in registers.

TESTED=unit

Change-Id: I360651ce1fdb1a792b174185f344c94493e4aea6