common: Refine the memory ambiguation optimization.
The current pass simply checks if the GEP of a store has ANY
non-inductive nodes, and if so, it serializes all stores originating
from that instruction. This results in overserialization when it is
actually provable that store addresses in a loop body are independent,
such as in the following example:

  unsigned n, m, x;  // variables only known at runtime, x < n.
  int array[n][m];
  for (int i = 0; i < m; i++) {
    array[x][i] = /* some value */;
  }

Since the size of the array is not statically known, the address
calculation is noninductive, and all stores to array become serialized,
even though it is obvious that those stores go to separate addresses.

This new pass identifies all the sources of the address calculation and
compares them across all the dynamic store nodes. If they only differ by
induction variables (i.e. share a base offset), then they will not
become serialized.

TESTED=unit

Change-Id: I9092436b8f4e25243ac431752a631f1ee1a2dc6d