common: Use different BGL remove_edge() overload.
Despite OutEdgeListS = boost::setS, remove_edge(edge_descriptor e) is
actually O(E/V); there is a linear search through the IN-edge list of
the target to find all edges with a matching EdgeProperty. This can
become a severe performance bottleneck on nodes with very high degree.

We don't allow our DDDG to be a multi-graph (no multiple edges with the
same source and target), so instead, we can use the API remove_edge(u,
v, graph) instead; this removes ALL edges between vertices u and v
(which is guaranteed to be one), and it also does not perform a linear
scan of the in-edge list.

Change-Id: Ib399373de118bc1967de4a7b9542cfd5221fed4c