all: Add support for vector operations and memory accesses.
The benefits of vector operations are obvious, not least because we can
bypass loop unrolling within the vector size and increase memory
bandwidth without partitioning. For ACP/cache accesses, the benefits are
even greater, since we now are not restricted to 32/64 bit loads and
stores.

We support GCC vector literals. An integration test has been added to
show how to write SIMD code, but the general gist is this:

  typedef int v8si __attribute__((__vector_size__(32)));
  v8si myvector = (v8si){0,1,2,3,4,5,6,7};

This creates a vector of 8 signed integers (total 32 bytes) and fills
them with values from 0 to 7.

SIMD support requires the recent update to LLVM-Tracer to properly write
vector data in the trace.

Change-Id: Idb733225d4239cf9f3b12fe96ef0fc2491eb4601