gem5: Use a tiny L1 cache to implement ACP coherency logic.
There are three ways to implement ACP:
  1. Add a special port to the gem5 Cache model itself to do exactly
     what ACP does. This is a lot of work.
  2. Add a port to the accelerator, connect it directly to the L2 cache
     and handle coherency from within the accelerator (mainly ownership
     requests on stores). This is the current implementation.
  3. Add a port to the accelerator, attach to a tiny L1 cache sitting
     between it at the L2 cache, and issue normal reads and writes. The
     L1 cache will handle coherency automatically.

The drawback of the current implementation is that the request for
ownership is an entirely separate transaction from the store itself, so
we have to pay for the latency of going to L2 twice on the critical
path. In other words, we ironically get less control of the desired
latency by doing it ourselves.

This patch replaces the existing implementation with method 3 as the
default (this can be changed dynamically).  With method 3, we can set
the tag/data/response latency of the L1 cache to be whatever we like, so
that a load or store to it looks like we're accessing L2. We make the
cache only one or two cache lines large so that we don't actually
introduce any significant cache space to the system. The drawback is
that depending on the workload, we might significantly decrease total L2
accesses, so anything that depends on that statistic (like a power
model) needs to be corrected for.

Change-Id: I31bf9d6e7994941c2021b0134fb2d112ceb18f22