common: In ready mode, dmaLoads should not get cleaned.
cleanLeafNodes is designed to identify and remove nodes from the graph
when we recognize that their produced values are never used and they have
no side effects. In ready mode, dmaLoad nodes do not have memory
dependences on other loads and stores since this relationship gets
captured with the full/empty bit mechanism, so for a dmaLoad to have no
children is a reasonable scenario.

Added documentation for this function.

Change-Id: I7205065780dc67706961d01feaee1e4845d66a80