common: Don't allow arrays to have base address 0.
A base address of zero is impossible since that would mean accessing
memory location 0x0, and it would result in an error when that array was
accessed. Enforcing this requirement means that the user must be more
careful in specifying arrays in the Aladdin configuration file, as
unused arrays would result in triggering this error.

Change-Id: Ia0be816f5f049e2b3f4a5f41f4d348e133dace72