common: Allow function call arguments to be temporaries.
If a function is called with an argument that is expected to be named
pointer but is actually a temporary, then that temporary must be handled
differently from non-temporary arguments. In the past, that temporary
would not be added to the list of call arguments, therefore causing
mismatches in processing forwarded call arguments.

Example:
  void func(float* ptr0, float* ptr1);

  int main() {
    float* myptr = /* some value */;
    func(NULL, myptr);  <- NULL is a temporary.
  }

This patch will correctly associate "ptr0" with a placeholder variable
and "ptr1" with "myptr", whereas in the past, "myptr" would get
associated with "ptr0", thus causing errors in base address resolution.

Change-Id: Ia307c278f6b1136ab3ea02464029e6700b06069d