common: Wrap logical array accessors with exception thrower.
This is often the source of hard to find segfaults. When an array that
doesn't exist is accessed, map::operator[] will create a new entry in
the map rather than fail, and map::at() will throw a std exception that
doesn't convey any further helpful information. Now, throw an
AladdinException instead, which at the minimum will contain the array
name.

There are a lot of potential callsites, so just try/catch at the most
common ones.

Change-Id: I5fa8835242b1cd85030be951cf993d74a128057a