graph_opts: Fix a corner case with loop flattening.
The loop flattening optimization assumes that all compute ops with the
same line number as the branch can be removed by turning them into
"Move" operations. If they were already inductive operations
(IndexAdd) though, this does nothing for performance modeling, as
inductive operations have zero latency.

However, it does change the node from a compute-op to a non-compute-op.
This interferes with the memory ambiguation pass, because that IndexAdd
operation could have been an operand to the memory address generation.
Memory ambiguation works by searching for all the sources of a memory
operation, which must be load or compute operations.  By changing an
IndexAdd to a Move, it is no longer a compute-op, the memory
ambiguation pass fails to identify this inductive source, and then the
lack of any inductive sources causes stores which are provably
independent to be serialized.

Change-Id: I24ce03b5896881db15d9312bf991bb804bead2c9