graph_opts: Fix bug where phi nodes aren't actually removed.
In the case that we have the chain of phi->convert, currently
the pass doesn't actually check that the parent of the convert is a phi
(it checks if it's a convert), so we would add new edges to the phi node
when we actually want to remove all edges.

Furthermore, the pass starts from the last node and iterates upwards, so
if have phi->convert, then that phi node would be added to
checked_phi_nodes. As a result, it gets skipped later, even if it still
has other children to be removed.

Change-Id: Iadaa2baef6290228815d4f5cd8dfe336bfaa3910