common: Fix bug with counting registers used.
In particular, we need to skip return nodes, because they are a child of
potentially thousands of nodes, and each one of those will result in an
additional perceived register read even though a return node doesn't
read data.

We can also skip convert operations - they are mostly logical and often
don't actually change the bits of a value.

Change-Id: I1fde857b134d9d4a3ee51ab2b097aa237afbe383