common: Calculate register power based on bytes accessed.
Rather than assuming that every single bit in a register must switch on
a read or write access, we only count power for the words that are
accessed. This fixes power overestimation in large register files where
only a few words are accessed at a time. Naturally, this assumes the
availability of word-level clock-gating.

Change-Id: Ibcc9210d00d4b13b33524ba018d83d657a19dda2