gem5: Add the ability to reset the trace back to the start.
This is useful if a program wants to loop a single trace over and over
again. Ordinarily, it would just stop when it gets to the end and never
build a DDDG, but now it can resume from the beginning simply by calling
resetTrace(accel_id) from within the simulated program.

Change-Id: I1e9114aa1dc7f08162077485ec5843a3e1417e46