all: Throw exceptions on invalid array accesses.
This covers cases of invalid addresses, invalid access sizes, and more.
This patch is not guaranteed to cover all possible places where the
exception should be caught.

This is part of the longer-term goal to remove as many assertion checks
based on invalid user input and replace them with exceptions that can
provide better debugging information.

Change-Id: I7223c6446aa537ff6226a9a5dcfbdad811b1f3a3