gem5: Prevent overlapping array address ranges.
Whenever we get a call to mapArrayToAccelerator(), we insert the
array name to base address and size mappings. It is possible for
another array to overlap with this address range, which is bad
because DMA nodes look up the array to use based on the base address,
and overlapping ranges create ambiguity.

To prevent this ambiguity, when we update an array mapping, we also
erase all previous array mappings that overlap with the new array range.
This is done simply by setting the size of the overlapping array range
to zero. The base address needs to be preserved, but the size will be
updated on the next call to mapArrayToAccelerator before it needs to be
used.

Change-Id: I6895c79dc17564414b61115fc34d7ae38431b340