common: Use a single opcode for all special math operations.
We had different opcodes for sin/cos operations, but there are many
other cmath operations. Adding a custom opcode for each of them is
bad. This uses a single opcode "MATH_OPS" for all the cmath operations.
Latency-wise, all the math operations currently take 3 cycles, which
is the latency we used for trig operations.

TODO: use different latencies for different special math functions.

Change-Id: I6991a18ce72c1609b389fbe8c645b6de0edb7928