gem5: Change due to recent merge in gem5 upstream.
In this commit, things listed below need to change in order to work
with the changes in upstream gem5.

1. Gabe Black renames the SimObject cxx_bases field to cxx_extra_bases
in commit b2d3cc0. As a result, we need to rename this field for
HybridDatapath as well.

2. Giacomo Travaglini changes the getMasterId function in commit
5187a24. HybridDatapath needs minor change in the constructor.

3. Giacomo Travaglini changes to use smart pointers for memory Requests
in commit f54020e. It completely drops the use of Raw Request pointers.
So all the newed Requests in HybridDatapath now become shared_ptrs. Also
the deletions on Requests are removed since those are not needed now.

4. Update the expected cycle numbers for two Ruby tests (slight variations).
All the CPU tests pass.
   a. MultipleInvocationsTest: 1212 -> 1244
   b. StreamingDmaTest: 514 -> 522

Change-Id: Id50266a44e4be402bb1179d3e3d8ac70d39b44c6