common: Add sampling for loops.
When the user is pretty confident that a loop is very repetitive
and likely to take the same latency for every iteration, then it's
a good candidate for doing sampling. This adds sampling to Aladdin.
An example shows below:

  int top_level(int a, int loop_size, int sample) {
    setSamplingFactor("loop", loop_size / sample);
    loop:
    // Instead of simulating all 10 iterations, sample only 1.
    for (int i = 0; i < sample; i++)
      a += i;
    return a;
  }

  int main() {
    int ret_val = top_level(1, 10, 1);
    printf("Result = %d\n", ret_val);
    return 0;
  }

The setSamplingFactor function sets up the sampled loop label and the
intended sampling factor. Aladdin will compute the latency for the
sampled loop and upsample it by the sampling factor. Sampling only
works for the outmost loop for now.

TESTED=unit,integration

Change-Id: I74d9717cfb62a9c8279f15867d3396edc8ba80b0