common: Fix the loop boundary when its next node is in a preheader.
The debug information generated by LLVM-6.0 can be inaccurate at
setting the line number of a preheader, which is crucial for Aladdin
to apply loop unrolling. As a heuristic, we will use the intermediate
branch before the preheader block as a loop boundary. In this case, to
get the target loop depth, we need to find the node after the preheader
block instead of the next node.

Change-Id: Ic944fb26fdf77960f9ca67f4abb11f01413e80d7