graph_opts: Fix loop pipelining across parent loop boundaries.
The loop pipelining pass takes all the iterations of a sampled loop, which will
erroneously pipeline across the outer loop’s boundaries. For example, the
following program has two loops, both have 2 iterations and we want to pipeline
loop1.

  int top_level(int a[2]) {
    int ret_val = 0;
    loop0:
    for (int i = 0; i < 2; i++) {
        loop1:
        for (int j = 0; j < 2; j++)
            ret_val += a[j];
    }
    return ret_val;
  }

The execution of the 4 loop0 iterations will look like this,

-----Cycle time----->
loop1-iter0
    loop1-iter1
        loop1-iter2
            loop1-iter3

However, because we don’t want to pipeline loop0, the timeline should be like
this,

-----Cycle time----->
loop1-iter0
    loop1-iter1
               loop1-iter2
                   loop1-iter3

To fix this, the solution first builds a tree the captures the hierarchical
structure of all loop iterations. And then it iterates over the tree to
apply pipelining transforms to the direct children of a parent loop.

TESTED=unit,integration

Change-Id: Iffb3b255e5dae54d871104c94bdf1b5173913d46