graph_opts: Fix when loop has fewer iterations than unrolling factor.
The issue is wrong loop boundary detection when a loop has fewer iterations
than its unrolling factor. In this case, the loop should be treated as if it's
flattened, and all the loop boundaries should be removed. The loop unrolling
pass has removed all the boundaries except the first one, which afterwards
confuses the loop tree creation procedure.

Change-Id: I91c33f754a0560f0a853e2046f7df6d920892025