sampling: Add support for sampling pipelined loops.
To upsample a pipelined loop, we need to calculate the average "termination interval"
between two adjacent iterations (so the loop should be sampled by at minimum two
iterations). As shown below, a termination interval is the part of the iteration
execution that can't be overlapped with the previous iteration. We can upsample
that interval to estimate the overall latency of the loop.

[-----iter0-----]
        [-----iter1-----]
                 *------*  <--- termination interval
               [-----iter2-----]
                        *------*  <--- termination interval

TESTED=unit

Change-Id: I22fcc20e21588aa21159029fbe2fce9b3a399138