graph_opts: Fix a bug in loop boundary detection.
If a loop has a preheader, the last instruction of the preheader will be a
branch that jumps into the loop, there will also be an intermediate branch that
jumps into the preheader block. The problem is that because these two branches
have the the same line number as the loop label, both will be identified as loop
boundaries (of which we only want the preheader branch).

The intermediate branch actually goes into the legacy code for handling a trace
that doesn't have any loop depth information and goto instructions. Since we now
rely on the loop depth information for loop boundary detection and goto is
notoriously bad (AFAIK, we don't have a use case for that), this commit removes
the legacy code and bypass the unrolling part (so that the intermediate branch
will not proceed).

NOTE: This is a breaking change and it requires the user to use the llvm-6.0
branch of LLVM-Tracer.

Change-Id: Id098f9dae79f4a344d1539a3ad79d073944598ba