gem5: Refactor code from HybridDatapath to Gem5Datapath.
This refactors out a few things from HybridDatapath to Gem5Datapath.

1) Three ports: DMA port, cache port and ACP port.
2) Callback functions used to update things in the associated datapath when the
   ports receive response (or retry, etc).
3) DMA request issue function. As the address translation is part of the request
   issue function, this also refactors out a translation function.
4) Use the override keyword when overriding a virtual function. This detects a
   bug in CachePort, as recvRespRetry is not a virtual function of MasterPort
   (so we don't need to implement that function).

Note the refactored DMA request issue function splits the request if it crosses
page boundaries. This was actually a bug in the Aladdin.

This also added a setParams() virtual function in Gem5Datapath. Any accelerator
inherited from Gem5Datapath should use this to set the parameters.

Change-Id: I6a374239065e2ee0841fc9d8fff7b827a604848f