graph_opts: Update the memory access type for host nodes.
We previously updated the host memory access type in the DDDG building process,
which can be incorrect if the host node's array label is not the original host
array label. To resolve that, this commit updates the host memory access type in
the DMA base address init pass, which traces back to the original node that has
the host array label.

Change-Id: Ied1c07389507fc8d85a136ef17f8df77083d04b2