graph_opts: Fix when loop iteration number is non-multiple of unrolling factor.
If the loop's iteration number is not multiples of the unrolling factor, the
exit branch should be a loop bound, otherwise we will have an isolated loop
bound that messes up the loop tree building process later.

This actually covers the case we previously handled, where the loop has fewer
iterations than the unrolling factor. This commit combines the two cases into
a general solution.

Change-Id: If5bdb8e5c37f5ad905ffa7d3361974449640b9a2