graph_opts: Fix a bug in finding the original variable/label of GEPs.
Previously in LLVM-3.4, when a multi-dimensional array is indexed, usually we
only see a single GEP instruction, as GEPs can have an arbitrary number of
indexes. But in LLVM-6, we are now observing cases where it gets split across
multiple GEPs. As a result, we need to find the original variable/label of
a GEP. An example is shown below, where a GEP feeding a host node has a parent
GEP.

The base address init pass finds the root parent of Load/Store ops in order to
get the real variable/label. To resolve this issue, this commit adds GEPs into
the pass so that it also finds the real variable/label for GEPs.

// GEP node.
0,-1,host_load_fp16,.lr.ph83:0,add.ptr15,32,59
2,64,0,1,idx.ext14,
1,64,0xebaae0,1,remote_data,
r,64,0xebaae0,1,add.ptr15,

...
// Another GEP node.
0,25,host_load_fp16,3:1,add.ptr17,32,69
2,64,0,1,4,
1,64,0xebaae0,1,add.ptr15,
r,64,0xebaae0,1,add.ptr17,

...
// Host node.
0,24,host_load_fp16,3:1,call19,107,72
4,64,0x621920,1,hostLoad,
1,64,0xec9280,1,add.ptr13,
2,64,0xebaae0,1,add.ptr17,
3,64,1568,1,conv18,
f,64,0xec9280,1,dst_addr,
f,64,0xebaae0,1,src_host_addr,
f,64,1568,1,size,
r,32,3,1,call19,

Change-Id: Ia81dc778b98da9510d5ba29fb6a5389404b0ac24