gem5: Add an accelerator-side command queue for buffering commands from CPUs.
Previously, it was undefined behaviour if the CPU tries to send commands to an
accelerator while it's still running. This adds a command queue in the
accelerator for buffering incoming commands from the CPUs, if it's busy. The
command queue obeys FIFO order. This is very useful if the CPU wants to invoke
multiple accelerators simutaneously. Without this, the CPU will have to wait if
it wants to invoke the same accelerator multiple times, blocking it from
further invoking other accelerators. With this, the inherent in-order invoking
of the same accelerator is ensured by the accelerator itself, freeing the CPU
from any blocking. The CPU only needs to check the corresponding finish flags
when it wants to.

TESTED=integration

Change-Id: Iefe95dccb9e112dc71f653e5d8789ca86508af49