gem5: Change to use an interrupt-based synchronization between CPU and accel.
We previously let the CPU poll on the finish flag allocated in shared memory for
synchronization. Polling leads to a lot of redundant memory transactions, which
is both inefficient software design and leads to slower simulations. This commit
changes that to an interrupt-based method, which suspends the CPU thread and the
accelerator will wake it up when it's finished, eliminating the polling-induced
events. Because spurious wakeups to the CPU by other system events (e.g., atomic
memory accesses) can occur, we still keep the finish flag and check it upon a
wakeup to ensure successful synchronization. Note that this method is not a true
hardware interrupt mechanism; instead, it relies on simulator hooks to wake up
CPU thread contexts that have been suspended.

Only a slight change is required in user programs. If the user wants a blocking
invocation, the usage is exactly the same as before. Otherwise if the
non-blocking API invokeAcceleratorAndReturn is used, instead of polling on the
finish flags as we used to, this adds a new API called waitForAccelerator
that uses the interrupt-based method as well.

All the integration tests are updated.

TESTED=integration

Change-Id: I7854df979039a9d151b73bca9ae3a3f62f55a228