all: Fix MemoryType dependency and bad include paths.
typedefs.h included gem5/aladdin_sys_connection.h just for the MemoryType enum
definition, and the include path goes above the root repo directory, which is
not a good idea, as it makes assumptions about what the repo directory is
actually named. This required splitting out the enum definition into a separate
header, as there was no other existing suitable place to put it.

Change-Id: Ic8521bc8b22d45a99c7731c2bb5c9b47baf211e3