unit-tests: Update test_loop_pipelining.
This tests per loop pipelining as opposed to global pipelining that it
previously tested. I remember I made some changes to the per loop pipelining
code in order to make it work with the loop sampling, but I didn't make the
similar changes to the global pipelining code, which is breaking the tests
currently. Let's check in this first, and I will go back to fix global
pipelining.

Change-Id: I8a3616a91fecd6260d38f346624c63aeb09a5acf