graph_opts: Fix a bug in memory ambiguation pass due to new LLVM version.
In this pass, we first find the GEP-store pairs for every dynamic store, and
then search the nodes that generate the source operands of every GEP in the
found pairs. For the same store instruction in consecutive loop iterations, if
the sources of the former is not independent of the latter, the two stores
should be serialized, and a memory edge will be added between the two nodes.

Previously in LLVM-3.4, when a multi-dimensional array is indexed, usually we
only see a single GEP instruction, as GEPs can have an arbitrary number of
indexes. But in LLVM-6, we are now observing cases where it gets split across
multiple GEPs. This causes the issue in this pass, where we used to only track
down the sources for one GEP. The fix is easy: if a GEP has a GEP parent, then
recursively track sources of its parent as well.

TESTED=unit

Change-Id: I1334132683e2fd0096b717cd8ee8f88801687c69