MachSuite: update benchmarks to DMA interface v3.
This requires allocating separate host/accelerator arrays, copying the
contents of bench_args_t into the host arrays and finally the result
back into the bench_args_t. We also use calloc() to zero out the
accelerator memory storing the output when possible, so we don't need to
dmaLoad a bunch of zeros. Finally, some arrays are no longer mapped to
the accelerator if they don't ever participate in host-to-accelerator
communication.

Follow-up commits can change the dmaLoad() calls to the helper methods
to pipeline DMA transactions in page sized chunks.

BUG=issue #15

TESTED=manually ran all benchmarks in native/tracer mode and gem5 and
verified output was correct.

Change-Id: Ieef084d77c8a5341b296b3eac66ec628a1b0236f