common: Enforce WAW dependencies for all nodes.
For whatever reason, the current code only enforced this between a
dmaLoad and a subsequent store and completely missed the more normal
store->store edge. This was hidden by the fact that most loops we write
don't store to the same address more than once, and loops act as control
flow blockers in Aladdin.

Change-Id: I877449f685a223ad3e4afe4e1201e1ce8814949b