unit-test: Improve Aladdin unit testing.
This is the first in hopefully a series of commits to clean up Aladdin
unit testing. Rather than comparing against a hardcoded number of nodes,
we either loosen the check (>= instead of ==) or instead check for the
presence of specific types of nodes using a set of Aladdin-specific
Catch-compatible unit testing matchers.

This commit adds matchers to test against connected nodes, so we can
assert that "this program contains 4 stores connected to this node"
with:

  REQUIRE_THAT(prog, ContainsStoreConnectedTo(node_id, 4))

Change-Id: I222e65225e89d77178c904457370467a752db35a