integration-test: Add automated MachSuite tests.
We previously had to manually use the sweep generator and run the
sweeps. This adds automated pytests for running MachSuite benchmarks.

This was initially written in unittest, but changed to pytest because
it offers a much easier way to enable dynamic test generation. Also,
pytest-xdist plugin is used to run the tests in parallel.

TESTED=integration-test

Change-Id: Ibf5e65b34512e46f2e05c68adc1308d5b4890ab0